+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Fixed
+
+- Alerts are now shown even if they're created from the web worker. In particular, bot exception alerts are now visible.
+
 ## [1.6.1] - 2024-11-16
 
 ### Added